Created separate ui components folder and abstracted table code
