+++ frontend/src/components/ShoppingList.tsx
@@ -1,9 +1,10 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-import { Card, Table } from "react-bootstrap";
+import { Card } from "react-bootstrap";
 import { SERVER_URL } from "./Config";
 import { URLS } from "./ConstantsPaths";
 import { getUserFromCookies, validateUserID } from "./UserAuthentication";
+import TableComponent from "ui_components/TableComponent";
 
 export const ShoppingList = () => {
   const [shoppinglist, setshoppinglist] = useState([]);
@@ -26,20 +27,10 @@ export const ShoppingList = () => {
     <Card className="mt-3">
       <Card.Body>
         <Card.Title>Shopping List</Card.Title>
-        <Table striped bordered hover>
-          <thead>
-            <tr>
-              <th>Ingredient</th>
-            </tr>
-          </thead>
-          <tbody>
-            {shoppinglist.map((item) => (
-              <tr key={item.id}>
-                <td>{item.ing.ingName}</td>
-              </tr>
-            ))}
-          </tbody>
-        </Table>
+        <TableComponent
+          headers={['Ingredient']}
+          rows={shoppinglist.map(item => [item.ing.ingName])}
+        />
       </Card.Body>
     </Card>
   );
