+++ frontend/src/ui_components/TableComponent.tsx
@@ -1,0 +1,24 @@
+import { Table } from 'react-bootstrap';
+
+export const TableComponent = ({ headers, rows }) => (
+    <Table striped bordered hover>
+        <thead>
+            <tr>
+                {headers.map((header, index) => (
+                    <th key={index}>{header}</th>
+                ))}
+            </tr>
+        </thead>
+        <tbody>
+            {rows.map((row, rowIndex) => (
+                <tr key={rowIndex}>
+                    {row.map((cell, cellIndex) => (
+                        <td key={cellIndex}>{cell}</td>
+                    ))}
+                </tr>
+            ))}
+        </tbody>
+    </Table>
+);
+
+export default TableComponent;
